pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "rocket"
version="0.0.1"
description = "ROCKET-1 Library"
dependencies = [
    "opencv-python==4.7.0.72",
    "setuptools>=49.2.0",
    "numpy",
    "einops",
    "requests>=2.20.0",
    "typing>=3.6.6",
    "lxml>=4.3.3",
    "psutil>=5.6.2",
    "Pyro4>=4.76",
    "coloredlogs>=10.0",
    "pillow>=8.0.0",
    "dill>=0.3.1.1",
    "daemoniker>=0.2.3",
    "xmltodict==0.12.0",
    "inflection>=0.3.1",
    "jinja2>=2.11.2",
    "imagehash>=4.0.0",
    "av>=11.0.0",
    "attr==0.3.2",
    "attrs==22.1.0",
    "dm_tree==0.1.7",
    "gym3==0.3.3",
    "gymnasium",
    "hydra_colorlog", 
    "hydra-core==1.3.2",
    "hydra-joblib-launcher",
    "Pillow==9.4.0",
    "transformers==4.31.0",
    "tree==0.2.4",
    "tabulate",
    "cchardet",
    "torch",
    "torchvision>=0.16.0",
    "torchaudio",
    "torchmetrics",
    "tokenizers",
    "timm==0.9.10", 
    "efficientnet_pytorch", 
    "lightning",
    "scipy",
    "flaky",
    "pyglet",
    "pynput",
    "rich",
    "ipdb",
    "watermark",
    "lz4",
    "kornia",
    "chardet", 
    "cuda-python",
    "diskcache", 
    "procgen",
    "cython==3.0.0a10",
    "ale-py",
    "humanize",
    "portalocker",
    "gradio==5.3.0", 
    "openai==1.52.2",
]
authors = [
    {name="Caishaofei", email="caishaofei@stu.pku.edu.cn"},
]
[options]
package_data = {'rocket'= ['*.json']}

[options.packages.find]
where = "rocket"

[tool.setuptools.packages.find]
include = ['rocket*']

[tool.setuptools.dynamic]
readme = {file = "README.md"}


[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
